Use trait in classes
