+++ inc/Engine/Optimization/GoogleFonts/AbstractGFOptimization.php
@@ -3,12 +3,16 @@ declare(strict_types=1);
 
 namespace WP_Rocket\Engine\Optimization\GoogleFonts;
 
+use WP_Rocket\Engine\Media\Fonts\FontsTrait;
+
 /**
  * Abstract Optimization Parent Class for Google Fonts Optimizers.
  *
  * @since 3.8
  */
 abstract class AbstractGFOptimization {
+	use FontsTrait;
+
 	/**
 	 * Allowed display values.
 	 *
@@ -129,54 +133,5 @@ abstract class AbstractGFOptimization {
 			'<link rel="preload" data-rocket-preload as="style" href="%1$s" /><link rel="stylesheet" href="%1$s" media="print" onload="this.media=\'all\'" /><noscript><link rel="stylesheet" href="%1$s" /></noscript>', // phpcs:ignore WordPress.WP.EnqueuedResources.NonEnqueuedStylesheet
 			$url
 		);
-	}
-
-	/**
-	 * Get the list of patterns to exclude from Google Fonts optimization which will have the be excluded from the combination.
-	 *
-	 * @return string[]
-	 */
-	protected function get_exclusions(): array {
-		/**
-		 * Filters the list of patterns to exclude from media font rewrite.
-		 *
-		 * @since 3.18
-		 *
-		 * @param string[] $exclusions The list of patterns to exclude from media fonts.
-		 */
-		return wpm_apply_filters_typed( 'string[]', 'rocket_exclude_locally_host_fonts', [] );
-	}
-
-		/**
-		 * Checks if a URL is excluded based on the provided exclusions.
-		 *
-		 * @since 3.18
-		 *
-		 * @param string   $url        The URL to check.
-		 * @param string[] $exclusions The list of exclusions.
-		 *
-		 * @return bool True if the URL is excluded, false otherwise.
-		 */
-	protected function is_excluded( string $url, array $exclusions ): bool {
-		if ( empty( $exclusions ) ) {
-			return false;
-		}
-
-		// Escape each exclusion pattern and combine them into a single regex pattern.
-		$escaped_exclusions = array_map(
-				function ( $exclusion ) {
-					return preg_quote( $exclusion, '#' );
-				},
-			$exclusions
-			);
-
-		$exclusions_str = implode( '|', $escaped_exclusions );
-
-		// Use a single regex pattern to check for any exclusion.
-		if ( preg_match( '#(' . $exclusions_str . ')#', $url ) ) {
-			return true;
-		}
-
-		return false;
 	}
 }
